Remove v8 and Refactoring
- Removed (almost) all references to v8 within the project :(
- Moved all the things into the Gunship namespace.
- Renamed Gunship::Gunship to Gunship::Engine
    - Gunship.h/.cpp became Engine.h/.cpp, and a new Gunship.h/.cpp was
added.
- Removed all existing components (everything in Components directory)
to make room for entityx.
