+++ include/Input.h
@@ -1,29 +1,32 @@
 # pragma once
 
-class Gunship;
-
-struct Input
+namespace Gunship
 {
-	std::vector< SDL_Scancode > downKeys;
-	std::vector< SDL_GameController* > controllers;
+	class Engine;
 
-	std::vector< SDL_Scancode > keyDownEvents;
-	std::vector< SDL_Scancode > keyUpEvents;
-	std::vector< SDL_JoyAxisEvent > joyAxisEvents;
+	struct Input
+	{
+		std::vector< SDL_Scancode > downKeys;
+		std::vector< SDL_GameController* > controllers;
 
-	bool exit;
+		std::vector< SDL_Scancode > keyDownEvents;
+		std::vector< SDL_Scancode > keyUpEvents;
+		std::vector< SDL_JoyAxisEvent > joyAxisEvents;
 
-	Input();
+		bool exit;
 
-	void ConsumeInput();
-	void Update( Gunship* gunship );
+		Input();
 
-	bool KeyPressed( SDL_Scancode key ) const;
-	bool KeyReleased( SDL_Scancode key ) const;
-	bool KeyUp( SDL_Scancode key ) const;
-	bool KeyDown( SDL_Scancode key ) const;
+		void ConsumeInput();
+		void Update( Engine* gunship );
 
-	float AxisMotion( Uint8 controller, Uint8 axis ) const;
-	float AxisValue( int controller, SDL_GameControllerAxis axis ) const;
-	float AxisValue( SDL_GameController*, SDL_GameControllerAxis axis ) const;
-};
+		bool KeyPressed( SDL_Scancode key ) const;
+		bool KeyReleased( SDL_Scancode key ) const;
+		bool KeyUp( SDL_Scancode key ) const;
+		bool KeyDown( SDL_Scancode key ) const;
+
+		float AxisMotion( Uint8 controller, Uint8 axis ) const;
+		float AxisValue( int controller, SDL_GameControllerAxis axis ) const;
+		float AxisValue( SDL_GameController*, SDL_GameControllerAxis axis ) const;
+	};
+}
